fix: correct HID report layout and MSRV, clear all clippy warnings
Two real defects surfaced by a clippy pass:

- KeyReport / MediaKeyReport are handed to the BLE stack as raw bytes via
  zerocopy's IntoBytes, but carried #[repr(packed)], which removes padding
  without pinning field order — Rust's field layout is unspecified without
  `C`. The HID report bytes happen to come out right today; a compiler that
  reorders would silently send garbage keystrokes. Now #[repr(C, packed)].

- rust-version claimed 1.77, but AtomicBool::fetch_not (app.rs) only
  stabilized in 1.81, so the declared MSRV never actually built. Bumped to
  1.81 to match reality.

The rest is mechanical: `cargo clippy --fix` over both feature sets, then
five `if let Err(_)` -> `.is_err()` by hand. The rotary-encoder branch was
restructured by lifting the if/else into a binding; the inner arms are
byte-identical, so the direction mapping is untouched.

Three crate-level allows are added with reasons rather than silently
suppressed: too_many_arguments (mode-entry functions take peripherals one
by one; reworking those signatures is its own task), upper_case_acronyms
(NEXT/AFE are hardware and esp-sr domain terms), and
field_reassign_with_default (ESP-IDF FFI config structs).

Three unused functions are marked #[allow(dead_code)] rather than deleted —
they look like a superseded JPEG path, but that is worth confirming before
removing.

Verified: clippy clean on both default and max2, fmt clean, both variants
build (2,922,752 B).

+++ src/lcd.rs
@@ -373,8 +373,8 @@ impl DisplayTargetDrive for FrameBuffer {
 
     fn flush(&mut self) -> anyhow::Result<()> {
         let bounding_box = self.bounding_box();
-        let x_start = bounding_box.top_left.x as i32;
-        let y_start = bounding_box.top_left.y as i32;
+        let x_start = bounding_box.top_left.x;
+        let y_start = bounding_box.top_left.y;
         let x_end = bounding_box.top_left.x + bounding_box.size.width as i32;
         let y_end = bounding_box.top_left.y + bounding_box.size.height as i32;
 
@@ -450,6 +450,7 @@ pub fn display_png<D: DisplayTargetDrive>(
     Ok(())
 }
 
+#[allow(dead_code)] // 旧 JPEG 路径,现走 new_jpg
 pub fn display_jpeg(jpeg: &[u8]) -> anyhow::Result<()> {
     let jpeg_buffer = crate::new_jpg::esp_jpeg_decode_one_picture(jpeg)?;
     log::info!(
@@ -584,7 +585,7 @@ fn new_terminal_renderer() -> embedded_graphics_terminal::TerminalRenderer {
 /// renderer 按 3 屏高创建,故 rows() 返回 3×可见行;sync_cells 与 parser 都用这个。
 pub fn terminal_text_cells() -> (u16, u16) {
     let renderer = new_terminal_renderer();
-    (renderer.cols() as u16, renderer.rows() as u16)
+    (renderer.cols(), renderer.rows())
 }
 
 impl UI {
